Sorting test and moving unique_values_from_queries to features/utilities
