+++ flowmachine/tests/test_unique_values_from_queries.py
@@ -4,9 +4,7 @@
 
 # -*- coding: utf-8 -*-
 
-import pytest
-
-from flowmachine.features.subscriber.unique_values_from_queries import (
+from flowmachine.features.utilities.unique_values_from_queries import (
     UniqueValuesFromQueries,
 )
 from flowmachine.features.utilities.events_tables_union import EventTableSubset
@@ -14,11 +12,12 @@ from flowmachine.features.utilities.events_tables_union import EventTableSubset
 
 def test_unique_subscribers_from_queries(get_dataframe):
     unique_subscriber_query = UniqueValuesFromQueries(
-        [EventTableSubset(start="2016-01-01", stop="2016-01-03")]
+        [EventTableSubset(start="2016-01-01", stop="2016-01-03")], ["imei", "imsi"]
     )
     # Using flowdb_synthetic_data (testdata seems to be empty?)
     print(get_dataframe(unique_subscriber_query))
-    assert (
-        get_dataframe(unique_subscriber_query).iloc[4]
-        == "7e3229b9b15ca4cf00f7ed0b494b2378"
-    )
+    assert get_dataframe(unique_subscriber_query).iloc[4].tolist() == [
+        "c38c773ae151c0573209c7697f6ba24a",
+        "52cb2c1a142e1c86b8f79d8783535fb8",
+    ]
+    assert get_dataframe(unique_subscriber_query).columns.tolist() == ["imei", "imsi"]
